+++ packages/fast-tooling-react/src/form/form-control-switch.tsx
@@ -6,6 +6,9 @@ import { FormChildOptionItem } from "./form.props";
 import { generateExampleData, isConst, isSelect } from "./utilities";
 import { ItemConstraints } from "./controls/control.array.props";
 import { SingleLineControlPlugin } from "./templates/plugin.control.single-line";
+import ControlPluginUtilities, {
+    ControlPluginUtilitiesProps,
+} from "./templates/plugin.control.utilities";
 
 class FormControlSwitch extends React.Component<FormControlSwitchProps, {}> {
     public static displayName: string = "FormControlSwitch";
@@ -22,45 +25,75 @@ class FormControlSwitch extends React.Component<FormControlSwitchProps, {}> {
      * Renders form items
      */
     private renderControl(): React.ReactNode {
+        let control: ControlPluginUtilities<ControlPluginUtilitiesProps>;
+
+        // Check to see if there is any associated `formControlId`
+        // then check for the id within the passed controlPlugins
+        if (typeof this.props.schema.formControlId === "string") {
+            control = this.props.controlPlugins.find(
+                (controlPlugin: StandardControlPlugin) => {
+                    return controlPlugin.matchesId(this.props.schema.formControlId);
+                }
+            );
+        }
+
         const hasEnum: boolean = isSelect({ enum: this.props.schema.enum });
 
         if (
             isConst(this.props.schema) ||
             (hasEnum && this.props.schema.enum.length === 1)
         ) {
-            return this.renderDisplay();
+            return this.renderDisplay(
+                control !== undefined ? control : this.props.controls.display
+            );
         }
 
         if (hasEnum) {
-            return this.renderSelect();
+            return this.renderSelect(
+                control !== undefined ? control : this.props.controls.select
+            );
         }
 
         if (this.props.schema.oneOf || this.props.schema.anyOf) {
-            return this.renderSectionLink();
+            return this.renderSectionLink(
+                control !== undefined ? control : this.props.controls.sectionLink
+            );
         }
 
         switch (this.props.schema.type) {
             case "boolean":
-                return this.renderCheckbox();
+                return this.renderCheckbox(
+                    control !== undefined ? control : this.props.controls.checkbox
+                );
             case "number":
-                return this.renderNumberField();
+                return this.renderNumberField(
+                    control !== undefined ? control : this.props.controls.numberField
+                );
             case "string":
-                return this.renderTextarea();
+                return this.renderTextarea(
+                    control !== undefined ? control : this.props.controls.textarea
+                );
             case "array":
-                return this.renderArray();
+                return this.renderArray(
+                    control !== undefined ? control : this.props.controls.array
+                );
             case "children":
-                return this.renderChildren();
+                return this.renderChildren(
+                    control !== undefined ? control : this.props.controls.children
+                );
             case "null":
-                return this.renderButton();
+                return this.renderButton(
+                    control !== undefined ? control : this.props.controls.button
+                );
             default:
-                return this.renderSectionLink();
+                return this.renderSectionLink(
+                    control !== undefined ? control : this.props.controls.sectionLink
+                );
         }
     }
 
-    private renderChildren(): React.ReactNode {
-        const standardControl: StandardControlPlugin = this.props.controls.children;
-
-        standardControl.updateProps(
+    private renderChildren(control: StandardControlPlugin): React.ReactNode {
+        control.updateProps(
             Object.assign(
                 {
                     childOptions: this.getChildOptions(),
@@ -70,16 +103,14 @@ class FormControlSwitch extends React.Component<FormControlSwitchProps, {}> {
             )
         );
 
-        return standardControl.render();
+        return control.render();
     }
 
     /**
      * Renders the array form item
      */
-    private renderArray(): React.ReactNode {
-        const standardControl: StandardControlPlugin = this.props.controls.array;
-
-        standardControl.updateProps(
+    private renderArray(control: StandardControlPlugin): React.ReactNode {
+        control.updateProps(
             Object.assign(
                 {
                     minItems: get(this.props.schema, ItemConstraints.minItems, 0),
@@ -90,16 +121,14 @@ class FormControlSwitch extends React.Component<FormControlSwitchProps, {}> {
             )
         );
 
-        return standardControl.render();
+        return control.render();
     }
 
     /**
      * Renders the number field form item
      */
-    private renderNumberField(): React.ReactNode {
-        const standardControl: StandardControlPlugin = this.props.controls.numberField;
-
-        standardControl.updateProps(
+    private renderNumberField(control: StandardControlPlugin): React.ReactNode {
+        control.updateProps(
             Object.assign(
                 {
                     min: this.props.schema.minimum,
@@ -110,41 +139,35 @@ class FormControlSwitch extends React.Component<FormControlSwitchProps, {}> {
             )
         );
 
-        return standardControl.render();
+        return control.render();
     }
 
     /**
      * Renders the checkbox form item
      */
-    private renderCheckbox(): React.ReactNode {
-        const singleLineControl: SingleLineControlPlugin = this.props.controls.checkbox;
+    private renderCheckbox(control: SingleLineControlPlugin): React.ReactNode {
+        control.updateProps(this.getCommonControlProps());
 
-        singleLineControl.updateProps(this.getCommonControlProps());
-
-        return singleLineControl.render();
+        return control.render();
     }
 
     /**
      * Renders a section link for properties
      * that are objects
      */
-    private renderSectionLink(): React.ReactNode {
-        const standardControl: StandardControlPlugin = this.props.controls.sectionLink;
+    private renderSectionLink(control: StandardControlPlugin): React.ReactNode {
+        control.updateProps(this.getCommonControlProps());
 
-        standardControl.updateProps(this.getCommonControlProps());
-
-        return standardControl.render();
+        return control.render();
     }
 
     /**
      * Renders the textarea form item
      */
-    private renderTextarea(): React.ReactNode {
+    private renderTextarea(control: StandardControlPlugin): React.ReactNode {
         const rows: number | undefined = this.props.schema.rows || void 0;
 
-        const standardControl: StandardControlPlugin = this.props.controls.textarea;
-
-        standardControl.updateProps(
+        control.updateProps(
             Object.assign(
                 {
                     rows,
@@ -153,22 +176,20 @@ class FormControlSwitch extends React.Component<FormControlSwitchProps, {}> {
             )
         );
 
-        return standardControl.render();
+        return control.render();
     }
 
     /**
      * Renders the select form item
      */
-    private renderSelect(): React.ReactNode {
+    private renderSelect(control: StandardControlPlugin): React.ReactNode {
         const options: any[] = this.props.schema.enum || [];
 
         if (!this.props.required && typeof options[0] !== "undefined") {
             options.unshift(void 0);
         }
 
-        const standardControl: StandardControlPlugin = this.props.controls.select;
-
-        standardControl.updateProps(
+        control.updateProps(
             Object.assign(
                 {
                     options,
@@ -177,26 +198,22 @@ class FormControlSwitch extends React.Component<FormControlSwitchProps, {}> {
             )
         );
 
-        return standardControl.render();
+        return control.render();
     }
 
     /**
      * Renders the display form item
      */
-    private renderDisplay(): React.ReactNode {
-        const standardControl: StandardControlPlugin = this.props.controls.display;
+    private renderDisplay(control: StandardControlPlugin): React.ReactNode {
+        control.updateProps(this.getCommonControlProps());
 
-        standardControl.updateProps(this.getCommonControlProps());
-
-        return standardControl.render();
+        return control.render();
     }
 
-    private renderButton(): React.ReactNode {
-        const standardControl: StandardControlPlugin = this.props.controls.button;
+    private renderButton(control: StandardControlPlugin): React.ReactNode {
+        control.updateProps(this.getCommonControlProps());
 
-        standardControl.updateProps(this.getCommonControlProps());
-
-        return standardControl.render();
+        return control.render();
     }
 
     private handleAddExampleData = (additionalSchemaPathSyntax: string): any => {