+++ packages/fast-tooling-react/src/form/templates/plugin.control.single-line.tsx
@@ -4,11 +4,8 @@ import ControlPluginUtilities, {
 } from "./plugin.control.utilities";
 import SingleLineControlTemplate from "./template.control.single-line";
 
-/* tslint:disable */
-export interface SingleLineControlPluginProps extends ControlPluginUtilitiesProps {}
-
 export class SingleLineControlPlugin extends ControlPluginUtilities<
-    SingleLineControlPluginProps
+    ControlPluginUtilitiesProps
 > {
     public render(): React.ReactNode {
         return <SingleLineControlTemplate {...this.props} {...this.config} />;