+++ packages/fast-color-explorer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.2](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-color-explorer@1.4.1...@microsoft/fast-color-explorer@1.4.2) (2019-10-07)
+
+**Note:** Version bump only for package @microsoft/fast-color-explorer
+
+
+
+
+
 ## [1.4.1](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-color-explorer@1.4.0...@microsoft/fast-color-explorer@1.4.1) (2019-10-02)
 
 **Note:** Version bump only for package @microsoft/fast-color-explorer